% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-ops.R
\name{add}
\alias{add}
\title{Add Two Numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A `numeric`.}

\item{y}{A `numeric`.}
}
\value{
Sum of `x` and `y`.
}
\description{
Takes two numbers and returns their sum
}
\examples{
add(1, 1)
add(2, 4)

\dontrun{
add(6, 12341) # Avoid running expensive calls
}
}
