% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-values.R
\name{count_na}
\alias{count_na}
\title{Count the number of missing values in a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The number of missing values in the vector
}
\description{
This function counts the number of missing values in a vector.
}
\examples{
# Some missing values
count_na(c(1, 2, NA, 4))

# All missing values
count_na(c(NA, NA, NA))

# No missing values
count_na(c(1, 2, 3))
}
