% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-ops.R
\name{divide}
\alias{divide}
\title{Divide Two Numbers}
\usage{
divide(x, y)
}
\arguments{
\item{x}{A `numeric`.}

\item{y}{A `numeric`.}
}
\value{
Quotient of `x` and `y`.
}
\description{
Takes two numbers and returns their quotient
}
\examples{
divide(1, 1)
divide(2, 4)

\dontrun{
divide(6, 12341) # Avoid running expensive calls
}
}
