% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-ops.R
\name{multiply}
\alias{multiply}
\title{Multiple Two Numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{A `numeric`.}

\item{y}{A `numeric`.}
}
\value{
Product of `x` and `y`.
}
\description{
Takes two numbers and returns their product
}
\examples{
multiply(1, 1)
multiply(4, 6)

\dontrun{
multiple(2, 311234) # Avoid running expensive calls
}
}
