% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netid-package.R
\docType{package}
\name{netid-package}
\alias{netid}
\alias{netid-package}
\title{netid: STATS 290 Example Package}
\description{
This is an example of an R package we created during class.
}
\author{
\strong{Maintainer}: J J B \email{j@somewhere.com} (\href{https://orcid.org/your-orcid-id-here}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Billy Joel (\href{https://orcid.org/their-orcid-id-here}{ORCID}) [contributor, medium]
}

}
\keyword{internal}
