% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-ops.R
\name{subtract}
\alias{subtract}
\title{Subtract Two Numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A `numeric`.}

\item{y}{A `numeric`.}
}
\value{
Difference of `x` and `y`.
}
\description{
Takes two numbers and returns their difference
}
\examples{
subtract(1, 1)
subtract(2, 4)

\dontrun{
subtract(6, 12341) # Avoid running expensive calls
}
}
