% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-stats290.R
\name{theme_stats290}
\alias{theme_stats290}
\title{Theme for STATS 290 Graphs}
\usage{
theme_stats290(...)
}
\arguments{
\item{...}{additional element specifications not
controlled by the theme.}
}
\value{
A ggplot2 theme object customized for STATS 290.
}
\description{
Apply the STATS 290 theme for ggplot2 graphs
}
\examples{
library(ggplot2)
ggplot(mtcars) +
   aes(x = wt, y = mpg,
       colour = factor(gear)) +
   geom_point() +
   theme_stats290() # Our theme
}
